PLA-6199 pull out paging
